Move all common tests to src
