+++ packages/common/package.json
@@ -11,7 +11,7 @@
     "format:fix": "prettier --write .",
     "format": "prettier --check .",
     "lint:fix": "yarn lint --fix",
-    "lint": "eslint --ext .ts --max-warnings 0 src test",
+    "lint": "eslint --ext .ts --max-warnings 0 src",
     "test": "mocha",
     "typecheck": "tsc --noEmit"
   },
